cstore: implement UnionDatapackStore.getdeltachain()
Summary:
Implements the getdeltachain function on the new UnionDatapackStore class.

This required some modifications to the DeltaChainIterator. Since the results of
the iterator may cross multiple different chains, we need to keep each chain
alive until the iterator is destructed, so we need to keep a reference to each
chain. We also had to remove the size() property from the iterator since the
fact that the chain spans chains means we don't know the size up front.

Test Plan: Adds a test

Reviewers: #mercurial, simonfar

Reviewed By: simonfar

Subscribers: simonfar, mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D4556458

Signature: t1:4556458:1487199872:07dffa3121acfbeb6d6993b518e6f4887122d4d5

+++ cstore/py-treemanifest.h
@@ -329,9 +329,9 @@ static PyObject *treemanifest_set(PyObject *o, PyObject *args) {
     // this is a remove operation!!
     self->tm.remove(std::string(filename, (size_t) filenamelen));
     Py_RETURN_NONE;
-  } else if (hashlen != BIN_NODE_SIZE) {
+  } else if (hashlen != (ssize_t)BIN_NODE_SIZE) {
     PyErr_Format(PyExc_ValueError,
-        "hash length must be %d bytes long", BIN_NODE_SIZE);
+        "hash length must be %zu bytes long", BIN_NODE_SIZE);
     return NULL;
   } else if (flagstrlen > 1) {
     PyErr_Format(PyExc_ValueError,
@@ -666,7 +666,7 @@ static int treemanifest_setitem(py_treemanifest *self, PyObject *key, PyObject *
     return -1;
   }
 
-  if (nodelen != BIN_NODE_SIZE) {
+  if (nodelen != (ssize_t)BIN_NODE_SIZE) {
       PyErr_Format(PyExc_ValueError, "invalid node length %zd", nodelen);
       return -1;
   }
